refator: extend class
