+++ testcafe/tests/drive/file_sharing_scenario.js
@@ -6,12 +6,14 @@ import {
   setDownloadPath,
   TESTCAFE_DRIVE_URL
 } from '../helpers/utils'
-import DrivePage from '../pages/drive-model'
+import PrivateDrivePage from '../pages/drive-model-private'
 import PublicDrivePage from '../pages/drive-model-public'
+import PublicViewerPage from '../pages/drive-viewer-model-public'
 
 let data = require('../helpers/data')
-const drivePage = new DrivePage()
+const privateDrivePage = new PrivateDrivePage()
 const publicDrivePage = new PublicDrivePage()
+const publicViewerPage = new PublicViewerPage()
 
 //************************
 //Tests when authentified
@@ -20,14 +22,14 @@ fixture`File link Sharing Scenario`.page`${TESTCAFE_DRIVE_URL}/`.beforeEach(
   async t => {
     console.group(`\n↳ ℹ️  Loggin & Initialization`)
     await t.useRole(driveUser)
-    await drivePage.waitForLoading()
+    await privateDrivePage.waitForLoading()
     console.groupEnd()
   }
 )
 
 test('Drive : Create a $test_date_time folder in Drive', async () => {
   console.group(`↳ ℹ️  Drive : Create a ${data.FOLDER_DATE_TIME} folder`)
-  await drivePage.addNewFolder(data.FOLDER_DATE_TIME)
+  await privateDrivePage.addNewFolder(data.FOLDER_DATE_TIME)
   //We need to pass data.FOLDER_DATE_TIME through multiple fixture, so we cannot use ctx here.
   console.groupEnd()
 })
@@ -38,14 +40,18 @@ test('Drive : from Drive, go in a folder, upload a file, and share the file', as
       data.FOLDER_DATE_TIME
     }, upload a file, and share the file`
   )
-  await drivePage.goToFolder(data.FOLDER_DATE_TIME)
-  await drivePage.openActionMenu()
+  await privateDrivePage.goToFolder(data.FOLDER_DATE_TIME)
+  await privateDrivePage.openActionMenu()
   await t.pressKey('esc') //close action Menu
 
-  await drivePage.uploadFiles([`${data.FILE_FROM_ZIP_PATH}/${data.FILE_XLSX}`])
-  await drivePage.shareFirstFilePublicLink()
+  await privateDrivePage.uploadFiles([
+    `${data.FILE_FROM_ZIP_PATH}/${data.FILE_XLSX}`
+  ])
+  await privateDrivePage.shareFirstFilePublicLink()
 
-  const link = await drivePage.copyBtnShareByLink.getAttribute('data-test-url')
+  const link = await privateDrivePage.copyBtnShareByLink.getAttribute(
+    'data-test-url'
+  )
   if (link) {
     data.sharingLink = link
     console.log(`SHARING_LINK : ` + data.sharingLink)
@@ -75,7 +81,7 @@ test(`[Desktop] Drive : Access a file public link, download the file, and check 
     `↳ ℹ️ [Desktop] Drive : Access a file public link, download the file, and check the 'create Cozy' link`
   )
   await t.navigateTo(data.sharingLink)
-  await publicDrivePage.waitForViewer()
+  await publicViewerPage.waitForLoading({ isSingleShareFile: true })
 
   await publicDrivePage.checkActionMenuPublicDesktop('file')
   await t
@@ -83,7 +89,7 @@ test(`[Desktop] Drive : Access a file public link, download the file, and check 
     .click(publicDrivePage.btnPublicDownload)
     .click(publicDrivePage.btnPublicCreateCozyFile)
   await publicDrivePage.checkCreateCozy()
-  await publicDrivePage.waitForViewer()
+  await publicViewerPage.waitForLoading({ isSingleShareFile: true })
 
   console.groupEnd()
 })
@@ -96,7 +102,7 @@ test(`[Mobile] Drive : Access a file public link, download the file, and check t
     portraitOrientation: true
   })
   await t.navigateTo(data.sharingLink)
-  await publicDrivePage.waitForViewer()
+  await publicViewerPage.waitForLoading({ isSingleShareFile: true })
 
   await publicDrivePage.checkActionMenuPublicMobile('file')
   await t
@@ -105,7 +111,7 @@ test(`[Mobile] Drive : Access a file public link, download the file, and check t
     .click(publicDrivePage.btnPublicMoreMenuFile) //need to re-open the more menu
     .click(publicDrivePage.btnPublicMobileCreateCozy)
   await publicDrivePage.checkCreateCozy()
-  await publicDrivePage.waitForViewer()
+  await publicViewerPage.waitForLoading({ isSingleShareFile: true })
 
   await t.maximizeWindow() //Back to desktop
   console.groupEnd()
@@ -118,15 +124,15 @@ fixture`Drive : Unshare public link`.page`${TESTCAFE_DRIVE_URL}/`.beforeEach(
   async t => {
     console.group(`\n↳ ℹ️  Loggin & Initialization`)
     await t.useRole(driveUser)
-    await drivePage.waitForLoading()
+    await privateDrivePage.waitForLoading()
     console.groupEnd()
   }
 )
 
 test('Unshare file', async () => {
   console.group('↳ ℹ️  Unshare file')
-  await drivePage.goToFolder(data.FOLDER_DATE_TIME)
-  await drivePage.unshareFirstFilePublicLink()
+  await privateDrivePage.goToFolder(data.FOLDER_DATE_TIME)
+  await privateDrivePage.unshareFirstFilePublicLink()
   console.groupEnd()
 })
 
@@ -144,7 +150,7 @@ test('Drive : No Access to an old file public link', async t => {
   console.group('↳ ℹ️  Drive : No Access to an old file public link')
   await t.navigateTo(data.sharingLink)
 
-  await publicDrivePage.waitForLoadingNotAvailable()
+  await publicDrivePage.waitForLoading({ isNotAvailable: true })
   await publicDrivePage.checkNotAvailable()
   console.groupEnd()
 })
@@ -156,14 +162,14 @@ fixture`Test clean up : remove files and folders`
   .page`${TESTCAFE_DRIVE_URL}/`.beforeEach(async t => {
   console.group(`\n↳ ℹ️  Loggin & Initialization`)
   await t.useRole(driveUser)
-  await drivePage.waitForLoading()
+  await privateDrivePage.waitForLoading()
   console.groupEnd()
 })
 
 test('(filesharing) Delete File, and foler', async () => {
   console.group('↳ ℹ️  Drive : Delete File, and foler')
-  await drivePage.goToFolder(data.FOLDER_DATE_TIME)
-  await drivePage.deleteElementByName(data.FILE_XLSX)
-  await drivePage.deleteCurrentFolder()
+  await privateDrivePage.goToFolder(data.FOLDER_DATE_TIME)
+  await privateDrivePage.deleteElementByName(data.FILE_XLSX)
+  await privateDrivePage.deleteCurrentFolder()
   console.groupEnd()
 })
